requirements.txt
streamlit
google-auth-oauthlib
google-api-python-client
deepl
python-dotenv
